Add ability to scan images and get JSON back
This is a very early prototype to prove the validility of using AWS
Rekognition to process an image and return labels associated with the
image.

We are essentially just doing a pass through here, where we take the
image and send it to AWS as-is and then get the JSON back.