Extract frames from video and perform image categorisation
This is a very rough-and-ready initial POC for taking a video file and
extracting frames per second for categorisation. It's extremely poor
performing, has had no review of memory management and doesn't do
anything with the JSON data returned from AWS.

This all being said, the initial testing is promising and suggests this
could very well be a viable option.